+++ src/utils/reattach.ts
@@ -55,5 +55,20 @@ export function mergeStates (
     }
   }
 
+    for (let i = 0, ids = Object.keys(comps), len = ids.length; i < len; i++) {
+      // replace component in contexts of spaces
+      comps[ids[i]].ctx.components = comps
+      if (!comps[ids[i]].isStatic) {
+        // replace outdated defs of dynamic components
+        let idParts = ids[i].split('$')
+        // is not root?
+        /* istanbul ignore else */
+        if (idParts.length > 1) {
+          let parentId = idParts.slice(0, -1).join('$')
+          comps[ids[i]].def = components[parentId].def.defs[comps[ids[i]].def.name]
+        }
+      }
+    }
+
   return comps
 }